requirements.txt
annotated-types==0.7.0
anyio==4.5.2
attrs==22.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.0.7
cryptography==44.0.0
cycler==0.11.0
distro==1.9.0
exceptiongroup==1.0.4
fastapi==0.115.8
fonttools==4.39.3
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib-resources==5.12.0
iniconfig==1.1.1
jiter==0.8.2
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.2
openai==1.63.2
packaging==22.0
pandas==1.5.3
pdfminer.six==20231228
pdfplumber==0.11.5
Pillow==9.5.0
pluggy==1.0.0
PyAudio==0.2.14
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.0.9
pypdfium2==4.30.1
pytest==7.2.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2022.7.1
PyYAML==6.0
requests==2.32.3
six==1.16.0
sniffio==1.3.1
SpeechRecognition==3.10.4
starlette==0.44.0
tomli==2.0.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.33.0
websockets==13.1
zipp==3.15.0

beautifulsoup4~=4.13.3
scikit-learn~=1.6.1